TX-only RAW socket: Avoid unneeded kernel packet_rcv() call
In case NFLOG is used for RX, then it is wasteful to also receive
(all) packets on the RAW socket, as it is only used for TX.
When RAW/AF_PACKET socket is only used for TX, tell the kernel
that we don't want to listen to any sll_protocol by passing zero.

This patch does not handle if the user doesn't use NFLOG, and
actually need RX on the RAW socket. I'll leave that up to Marek
to fix ;-)

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>